Pay out system implemented & fixed few bugs & improved code

+++ src/Domain/Ports/PayOut.php
@@ -6,5 +6,5 @@ use Ramsey\Uuid\UuidInterface;
 
 interface PayOut
 {
-    public function payOut(UuidInterface $walletId, int $amountOfMoney): void;
+    public function payOut(UuidInterface $walletId, float $amountOfMoney): void;
 }
